--- swanlab/data/run/db.py
@@ -1,23 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-r"""
-@DATE: 2024-01-19 15:46:28
-@File: swanlab/data/run/db.py
-@IDE: vscode
-@Description:
-    数据库连接串
-"""
-from ...db import (
-    Project,
-    Experiment,
-    Tag,
-    Chart,
-    Namespace,
-    Display,
-    Source,
-)
-from ...db import (
-    ExistedError,
-    NotExistedError,
-    ChartTypeError,
-)
