+++ swanlab/data/run/main.py
@@ -7,328 +7,23 @@ r"""
 @Description:
     在此处定义SwanLabRun类并导出
 """
-from typing import Any
 from ..settings import SwanDataSettings
-from ...log import swanlog
+from swanlab.log import swanlog
 from ..system import get_system_info, get_requirements
-from .utils import (
+from swanlab.utils.file import (
     check_exp_name_format,
-    check_desc_format,
-    get_a_lock,
-    json_serializable,
+    check_desc_format
 )
 from datetime import datetime
 import time
 import random
 import ujson
 from .exp import SwanLabExp
-from collections.abc import Mapping
-from .db import Experiment, ExistedError, NotExistedError
+from swanlab.db import Experiment, ExistedError, NotExistedError
 from typing import Tuple, Callable
-import yaml
-import argparse
-from ..modules import BaseType
+from swanlab.data.modules import BaseType
+from swanlab.data.config import SwanLabConfig
 from swanlab.cloud import ThreadPool
-
-
-def need_inited(func):
-    """装饰器，用于检查是否已经初始化"""
-
-    def wrapper(self, *args, **kwargs):
-        if not self._inited:
-            raise RuntimeError("You must call swanlab.init() before using swanlab.log")
-        return func(self, *args, **kwargs)
-
-    return wrapper
-
-
-class SwanLabConfig(Mapping):
-    """
-    The SwanConfig class is used for realize the invocation method of `run.config.lr`.
-    """
-
-    # 配置字典
-    __config = dict()
-
-    # 运行时设置
-    __settings = dict()
-
-    @property
-    def _inited(self):
-        return self.__settings.get("save_path") is not None
-
-    def __init__(self, config: dict = None, settings: SwanDataSettings = None):
-        """
-        实例化配置类，如果settings不为None，说明是通过swanlab.init调用的，否则是通过swanlab.config调用的
-
-        Parameters
-        ----------
-        settings : SwanDataSettings, optional
-            运行时设置
-        """
-        self.__config.update(self.__check_config(config))
-        self.__settings["save_path"] = settings.config_path if settings is not None else None
-        if self._inited:
-            self.__save()
-
-    @staticmethod
-    def __check_config(config: dict) -> dict:
-        """
-        检查配置是否合法，确保它可以被 JSON/YAML 序列化。
-        如果传入的是 argparse.Namespace 类型，会先转换为字典。
-        """
-        if config is None:
-            return {}
-        # config必须可以被json序列化
-        try:
-            if isinstance(config, argparse.Namespace):
-                config = vars(config)
-            # 将config转换为json序列化的dict
-            config = json_serializable(dict(config))
-            # 尝试序列化，如果还是失败就退出
-            yaml.dump(config)
-        except:
-            raise TypeError(f"config: {config} is not a valid dict, which can be json serialized")
-        return config
-
-    @staticmethod
-    def __check_private(name: str):
-        """
-        检查属性名是否是私有属性,如果是私有属性，抛出异常
-
-        Parameters
-        ----------
-        name : str
-            属性名
-
-        Raises
-        ----------
-        AttributeError
-            如果属性名是私有属性，抛出异常
-        """
-        methods = ["set", "get", "pop"]
-        swanlog.debug(f"Check private attribute: {name}")
-        if name.startswith("__") or name.startswith("_SwanLabConfig__") or name in methods:
-            raise AttributeError("You can not get private attribute")
-
-    @need_inited
-    def __setattr__(self, name: str, value: Any) -> None:
-        """
-        自定义属性设置方法。如果属性名不是私有属性，则同时更新配置字典并保存。
-        允许通过点号方式设置属性，但不允许设置私有属性：
-        ```python
-        run.config.lr = 0.01  # 允许
-        run.config._lr = 0.01 # 允许
-        run.config.__lr = 0.01 # 不允许
-        ```
-
-        值得注意的是类属性的设置不会触发此方法
-        """
-        # 判断是否是私有属性
-        self.__check_private(name)
-        # 设置属性，并判断是否已经初始化，如果是，则调用保存方法
-        self.__dict__[name] = value
-        # 同步到配置字典
-        self.__config[name] = value
-        self.__save()
-
-    @need_inited
-    def __setitem__(self, name: str, value: Any) -> None:
-        """
-        以字典方式设置配置项的值，并保存，但不允许设置私有属性：
-        ```python
-        run.config["lr"] = 0.01  # 允许
-        run.config["_lr"] = 0.01 # 允许
-        run.config["__lr"] = 0.01 # 不允许
-        ```
-        """
-        # 判断是否是私有属性
-        self.__check_private(name)
-        self.__config[name] = value
-        self.__save()
-
-    @need_inited
-    def set(self, name: str, value: Any) -> None:
-        """
-        Explicitly set the value of a configuration item and save it. For example:
-
-        ```python
-        run.config.set("lr", 0.01)   # Allowed
-        run.config.set("_lr", 0.01)  # Allowed
-        run.config.set("__lr", 0.01) # Not allowed
-        ```
-
-        Parameters
-        ----------
-        name: str
-            Name of the configuration item
-        value: Any
-            Value of the configuration item
-
-        Raises
-        ----------
-        AttributeError
-            If the attribute name is private, an exception is raised
-        """
-        self.__check_private(name)
-        self.__config[name] = value
-        self.__save()
-
-    @need_inited
-    def pop(self, name: str) -> bool:
-        """
-        Delete a configuration item; if the item does not exist, skip.
-
-        Parameters
-        ----------
-        name : str
-            Name of the configuration item
-
-        Returns
-        ----------
-        bool
-            True if deletion is successful, False otherwise
-        """
-        try:
-            del self.__config[name]
-            self.__save()
-            return True
-        except KeyError:
-            return False
-
-    @need_inited
-    def get(self, name: str):
-        """
-        Get the value of a configuration item. If the item does not exist, raise AttributeError.
-
-        Parameters
-        ----------
-        name : str
-            Name of the configuration item
-
-        Returns
-        ----------
-        value : Any
-            Value of the configuration item
-
-        Raises
-        ----------
-        AttributeError
-            If the configuration item does not exist, an AttributeError is raised
-        """
-        try:
-            return self.__config[name]
-        except KeyError:
-            raise AttributeError(f"You have not retrieved '{name}' in the config of the current experiment")
-
-    @need_inited
-    def update(self, data: dict):
-        """
-        Update the configuration item with the dict provided and save it.
-
-        Parameters
-        ----------
-        data : dict
-            Dict of the configuration item
-        """
-
-        self.__config.update(self.__check_config(data))
-        self.__save()
-
-    @need_inited
-    def __getattr__(self, name: str):
-        """
-        如果以点号方式访问属性且属性不存在于类中，尝试从配置字典中获取。
-        """
-        try:
-            return self.__config[name]
-        except KeyError:
-            raise AttributeError(f"You have not get '{name}' in the config of the current experiment")
-
-    @need_inited
-    def __getitem__(self, name: str):
-        """
-        以字典方式获取配置项的值。
-        """
-        try:
-            return self.__config[name]
-        except KeyError:
-            raise AttributeError(f"You have not get '{name}' in the config of the current experiment")
-
-    @need_inited
-    def __delattr__(self, name: str) -> bool:
-        """
-        删除配置项，如果配置项不存在,跳过
-
-        Parameters
-        ----------
-        name : str
-            配置项名称
-
-        Returns
-        ----------
-        bool
-            是否删除成功
-        """
-        try:
-            del self.__config[name]
-            return True
-        except KeyError:
-            return False
-
-    @need_inited
-    def __delitem__(self, name: str) -> bool:
-        """
-        删除配置项，如果配置项不存在,跳过
-
-        Parameters
-        ----------
-        name : str
-            配置项名称
-
-        Returns
-        ----------
-        bool
-            是否删除成功
-        """
-        try:
-            del self.__config[name]
-            return True
-        except KeyError:
-            return False
-
-    def __save(self):
-        """
-        保存config为json，不必校验config的YAML格式，将在写入时完成校验
-        """
-        swanlog.debug("Save config to {}".format(self.__settings.get("save_path")))
-        with get_a_lock(self.__settings.get("save_path"), "w") as f:
-            # 将config的每个key的value转换为desc和value两部分，value就是原来的value，desc是None
-            # 这样做的目的是为了在web界面中显示config的内容,desc是用于描述value的
-            config = {
-                key: {
-                    "desc": None,
-                    "sort": index,
-                    "value": value,
-                }
-                for index, (key, value) in enumerate(self.__config.items())
-            }
-            yaml.dump(config, f)
-
-    def __iter__(self):
-        """
-        返回配置字典的迭代器。
-        """
-        return iter(self.__config)
-
-    def __len__(self):
-        """
-        返回配置项的数量。
-        """
-        return len(self.__config)
-
-    def __str__(self):
-        return str(self.__config)
 
 
 class SwanLabRun:
@@ -375,7 +70,7 @@ class SwanLabRun:
         """
         # ---------------------------------- 初始化类内参数 ----------------------------------
         # 生成一个唯一的id，随机生成一个8位的16进制字符串，小写
-        _id = hex(random.randint(0, 2**32 - 1))[2:].zfill(8)
+        _id = hex(random.randint(0, 2 ** 32 - 1))[2:].zfill(8)
         timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
         self.__run_id = "run-{}-{}".format(timestamp, _id)
         # 初始化配置
